requirements.txt
opencv-python>=4.5.0
face-recognition>=1.3.0
numpy>=1.19.0
pandas>=1.2.0
Pillow>=8.0.0
openpyxl>=3.0.0
dlib>=19.22.0
cmake>=3.18.0
tkcalendar>=1.6.0
reportlab>=3.6.0
